refactor: group과 order 의존관계 역전
