+++ src/main/java/kitchenpos/table/application/Validator.java
@@ -1,8 +1,11 @@
 package kitchenpos.table.application;
 
+import java.util.List;
 import kitchenpos.table.domain.OrderTable;
 
 public interface Validator {
 
     void validateOrderStatus(final OrderTable orderTable);
+
+    void validateOrdersStatus(final List<OrderTable> orderTables);
 }
